[nodejs] Aligns development and CI flows to avoid issues (#67)
Fixes ES-492.

This PR aligns the development an CI flows when it comes to where they
expect the forked dependencies to be. If they are in different places
between development and CI flows, there might be some issues with the CI
build.

This happens because even though the forked libraries might be put
wherever we want at development time, in the CI they are expected to be
inside the root folder of the repository because of limitations in the
CI runners.

The problem starts because the location of the libraries during
development will influence the contents of the `package.json` files,
which will be wrong from the perspective of the CI.

**tl;dr:** all the forked repos have to be cloned **inside** the root
folder of this repository. This should **not** be changed.

Additionally the following was done:

- Forked libraries added to `.gitignore`.
- `.envrc` (auto env-var loader) added to `.gitignore`.
- `mise.toml` (polyglot version manager configuration file) added to
`.gitignore`.
- CI runner now uses the `setup-nodejs` action.
- CI runner nodejs was upgraded to version 20 to clear up CI warnings at
build time.
- CI runner OS upgraded to Ubuntu 24.04.
- The layer build script runs an `npm cache clean --force` before doing
anything to ensure a clean starting state.
- The README has been updated with better development instructions
according to recent changes and experiences.
- `opentelemetry-propagator-aws-xray` and
`opentelemetry-propagator-aws-xray-lambda` are manually compiled, as
they are new dependencies of `opentelemetry-instrumentation-aws-lambda`.